requirements.txt
joblib==1.5.2
numpy==2.3.3
pillow==11.3.0
scikit-learn==1.7.2
scipy==1.16.2
threadpoolctl==3.6.0
